refactor: componentize client list and fix clients service
