+++ src/app/admin/clientes/_components/client-rows/client-rows-props.tsx
@@ -1,0 +1,8 @@
+// Services
+import { type Client } from "@/app/services/clients";
+
+// DTOs
+export interface ClientRowsProps {
+  clients: Client[];
+  loading: boolean;
+}
