+++ src/app/admin/clientes/_components/client-table/client-table-props.tsx
@@ -1,0 +1,9 @@
+// Services
+import { type ClientListResponse } from "@/app/services/clients";
+
+// DTOs
+export interface ClientTableProps {
+  state: ClientListResponse;
+  loadPage: (page: number) => void;
+  loading: boolean;
+}
