+++ src/components/app-sidebar/app-sidebar-props.tsx
@@ -1,0 +1,8 @@
+// External libs
+import { type ComponentProps } from "react";
+
+// Components
+import { Sidebar } from "@/components/ui/sidebar";
+
+// DTOs
+export type AppSidebarProps = ComponentProps<typeof Sidebar>;
