+++ src/components/page-header/page-header-props.tsx
@@ -1,0 +1,8 @@
+// DTOs
+import { type ReactNode } from "react";
+
+export interface PageHeaderProps {
+  title: string;
+  description?: string;
+  children?: ReactNode;
+}
