+++ src/components/search-form/index.tsx
@@ -1,13 +1,18 @@
-import { Search } from "lucide-react"
+// External libs
+import { Search } from "lucide-react";
 
-import { Label } from "@/components/ui/label"
+// Components
+import { Label } from "@/components/ui/label";
 import {
   SidebarGroup,
   SidebarGroupContent,
   SidebarInput,
-} from "@/components/ui/sidebar"
+} from "@/components/ui/sidebar";
 
-export function SearchForm({ ...props }: React.ComponentProps<"form">) {
+// DTOs
+import { type SearchFormProps } from "./search-form-props";
+
+export function SearchForm({ ...props }: SearchFormProps) {
   return (
     <form {...props}>
       <SidebarGroup className="py-0">
@@ -20,9 +25,9 @@ export function SearchForm({ ...props }: React.ComponentProps<"form">) {
             placeholder="Search the docs..."
             className="pl-8"
           />
-          <Search className="pointer-events-none absolute top-1/2 left-2 size-4 -translate-y-1/2 opacity-50 select-none" />
+          <Search className="pointer-events-none absolute top-1/2 left-2 size-4 -translate-y-1/2 select-none opacity-50" />
         </SidebarGroupContent>
       </SidebarGroup>
     </form>
-  )
+  );
 }
