+++ src/components/search-form/search-form-props.tsx
@@ -1,0 +1,5 @@
+// External libs
+import { type ComponentProps } from "react";
+
+// DTOs
+export type SearchFormProps = ComponentProps<"form">;
