+++ src/components/version-switcher/version-switcher-props.tsx
@@ -1,0 +1,5 @@
+// DTOs
+export interface VersionSwitcherProps {
+  versions: string[];
+  defaultVersion: string;
+}
